+++ packages/extensions/link/src/index.ts
@@ -1,1 +1,8 @@
 export { LinkExtension, type LinkConfig } from "./extension";
+
+export {
+    $createLinkNode,
+    $isAutoLinkNode,
+    $isLinkNode,
+    TOGGLE_LINK_COMMAND,
+} from "@lexical/link";