+++ packages/utils/src/lexical/utils/index.ts
@@ -1,1 +1,2 @@
-export * from "@lexical/utils";
+
+export { $findMatchingParent, mergeRegister } from "@lexical/utils";